requirements.txt
aiohttp>=3.9.0
py-clob-client>=0.2.0
python-dotenv>=1.0.0
pytz>=2024.1
